# Optimize DB lookups for projects views