--- a/cadasta/organization/tests/test_views_api_projects.py
+++ b/cadasta/organization/tests/test_views_api_projects.py
@@ -1,4 +1,4 @@
-import json
+gimport json
 
 from django.test import TestCase
 from rest_framework.exceptions import PermissionDenied
@@ -503,6 +503,30 @@ class ProjectListAPITest(APITestCase, UserTestCase, TestCase):
         assert len(response.content['results']) == 3
         names = [proj['name'] for proj in response.content['results']]
         assert names == sorted(names, reverse=True)
+
+    def test_permission_filter(self):
+        addtional_clause = [{
+            'effect': 'allow',
+            'object': ['project/*/*'],
+            'action': ['party.create']
+        }, {
+            'effect': 'deny',
+            'object': ['project/*/unauthorized'],
+            'action': ['party.create']
+        }]
+
+        ProjectFactory.create_from_kwargs([
+            {'slug': 'unauthorized', 'organization': self.organization},
+            {'organization': self.organization}
+        ])
+
+        assign_policies(self.user, add_clauses=addtional_clause)
+
+        response = self.request(user=self.user,
+                                get_data={'permissions': 'party.create'})
+        assert response.status_code == 200
+        assert len(response.content['results']) == 1
+        assert response.content['results'][0]['slug'] != 'unauthorized'
 
     # CONDITIONS:
     #