--- a/cadasta/organization/tests/test_views_api_projects.py
+++ b/cadasta/organization/tests/test_views_api_projects.py
@@ -1,4 +1,4 @@
-gimport json
+import json
 
 from django.test import TestCase
 from rest_framework.exceptions import PermissionDenied