--- a/cadasta/organization/views/api.py
+++ b/cadasta/organization/views/api.py
@@ -187,14 +187,6 @@ class ProjectList(PermissionsFilterMixin,
                   mixins.ProjectQuerySetMixin,
                   generics.ListAPIView):
 
-    def permission_filter(self, view, p):
-        if p.archived is True:
-            return ('project.view_archived',)
-        elif p.access == 'private':
-            return ('project.view_private',)
-        else:
-            return ('project.view',)
-
     serializer_class = serializers.ProjectSerializer
     filter_backends = (filters.DjangoFilterBackend,
                        filters.SearchFilter,
@@ -202,8 +194,7 @@ class ProjectList(PermissionsFilterMixin,
     filter_fields = ('archived',)
     search_fields = ('name', 'organization__name', 'country', 'description',)
     ordering_fields = ('name', 'organization', 'country', 'description',)
-    permission_required = {'GET': 'project.list'}
-    permission_filter_queryset = permission_filter
+    permission_required = ()
 
 
 class ProjectDetail(APIPermissionRequiredMixin,