@alukach 2017-11-15
I'm a bit clueless as to how this PR has messed the tests up so badly for travis... anyone else seeing this on other branches? 

```python
py35-django-unit installed: argon2-cffi==16.3.0,awscli==1.11.167,beautifulsoup4==4.6.0,boto3==1.4.4,botocore==1.7.25,cadasta-platform==0.0.1,cadasta-test==0.2.7.1,certifi==2017.11.5,cffi==1.11.2,chardet==3.0.4,colorama==0.3.7,coverage==4.4.2,decorator==4.1.2,defusedxml==0.5.0,Django==1.10.8,django-allauth==0.33.0,django-appconf==1.0.2,django-audit-log==0.7.0,django-buckets==0.1.22,django-compressor==2.2,django-cors-headers==2.1.0,django-countries==5.0,django-crispy-forms==1.6.1,django-debug-toolbar==1.8,django-extensions==1.9.1,django-filter==1.0.4,django-formtools==2.1,django-jsonattrs==0.1.23,django-leaflet==0.18.1,django-parsley==0.6,django-sass-processor==0.5.5,django-simple-history==1.9.0,django-skivvy==0.1.9,django-tutelary==0.1.21,django-widget-tweaks==1.4.1,djangorestframework==3.6.4,djangorestframework-gis==0.11.2,djoser==0.7.0,docutils==0.14,drfdocs-cadasta==0.0.12,et-xmlfile==1.0.1,factory-boy==2.9.2,Faker==0.8.7,flake8==3.4.1,GDAL==1.10.0,gpxpy==1.1.2,idna==2.6,ipython==6.2.1,ipython-genutils==0.2.0,jdcal==1.3,jedi==0.11.0,jmespath==0.9.3,jsonschema==2.6.0,libsass==0.13.4,lxml==3.5.0,mccabe==0.6.1,numpy==1.13.3,oauthlib==2.0.6,olefile==0.44,openpyxl==2.4.8,pandas==0.20.3,parso==0.1.0,pexpect==4.3.0,pickleshare==0.7.4,Pillow==4.3.0,pluggy==0.5.2,prompt-toolkit==1.0.15,psycopg2==2.7.3.1,ptyprocess==0.5.2,py==1.5.0,pyasn1==0.3.7,pycodestyle==2.3.1,pycparser==2.18,pyflakes==1.5.0,Pygments==2.2.0,pylibmc==1.5.2,pyparsing==2.2.0,pytest==3.2.4,pytest-cov==2.5.1,pytest-django==3.1.2,python-dateutil==2.6.1,python-magic==0.4.13,python3-openid==3.1.0,pytz==2017.2,pyxform-cadasta==0.9.22,PyYAML==3.12,rcssmin==1.0.6,requests==2.18.4,requests-oauthlib==0.8.0,rfc3987==1.3.7,rjsmin==1.0.12,rsa==3.4.2,s3transfer==0.1.11,selenium==3.6.0,Shapely==1.6.1,simplegeneric==0.8.1,simplejson==3.11.1,six==1.11.0,sqlparse==0.2.4,text-unidecode==1.1,tox==2.9.1,traitlets==4.3.2,transifex-client==0.12.4,unicodecsv==0.14.1,urllib3==1.22,virtualenv==15.1.0,wcwidth==0.1.7,xlrd==0.9.4
py35-django-unit runtests: PYTHONHASHSEED='2487937216'
py35-django-unit runtests: commands[0] | ./runtests.py
Traceback (most recent call last):
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 510, in load_setuptools_entrypoints
    plugin = ep.load()
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/pkg_resources/__init__.py", line 2421, in load
    self.require(*args, **kwargs)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/pkg_resources/__init__.py", line 2444, in require
    items = working_set.resolve(reqs, env, installer, extras=self.extras)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/pkg_resources/__init__.py", line 875, in resolve
    raise VersionConflict(dist, req).with_context(dependent_req)
pkg_resources.ContextualVersionConflict: (py 1.5.0 (/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages), Requirement.parse('py<1.5,>=1.4.33'), {'pytest'})
During handling of the above exception, another exception occurred:
Traceback (most recent call last):
  File "/home/travis/build/Cadasta/cadasta-platform/runtests.py", line 155, in <module>
    exit_on_failure(pytest.main(pytest_args))
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/config.py", line 49, in main
    config = _prepareconfig(args, plugins)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/config.py", line 168, in _prepareconfig
    pluginmanager=pluginmanager, args=args)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 745, in __call__
    return self._hookexec(self, self._nonwrappers + self._wrappers, kwargs)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 339, in _hookexec
    return self._inner_hookexec(hook, methods, kwargs)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 334, in <lambda>
    _MultiCall(methods, kwargs, hook.spec_opts).execute()
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 613, in execute
    return _wrapped_call(hook_impl.function(*args), self.execute)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 250, in _wrapped_call
    wrap_controller.send(call_outcome)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/helpconfig.py", line 68, in pytest_cmdline_parse
    config = outcome.get_result()
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 279, in get_result
    raise ex[1].with_traceback(ex[2])
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 265, in __init__
    self.result = func()
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 614, in execute
    res = hook_impl.function(*args)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/config.py", line 957, in pytest_cmdline_parse
    self.parse(args)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/config.py", line 1121, in parse
    self._preparse(args, addopts=addopts)
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/config.py", line 1084, in _preparse
    self.pluginmanager.load_setuptools_entrypoints('pytest11')
  File "/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages/_pytest/vendored_packages/pluggy.py", line 515, in load_setuptools_entrypoints
    "Plugin %r could not be loaded: %s!" % (ep.name, e))
_pytest.vendored_packages.pluggy.PluginValidationError: Plugin 'django' could not be loaded: (py 1.5.0 (/home/travis/build/Cadasta/cadasta-platform/.tox/py35-django-unit/lib/python3.5/site-packages), Requirement.parse('py<1.5,>=1.4.33'), {'pytest'})!
ERROR: InvocationError: '/home/travis/build/Cadasta/cadasta-platform/runtests.py'
___________________________________ summary ____________________________________
ERROR:   py35-django-unit: commands failed
```