@seav 2017-11-15
> anyone else seeing this on other branches?

It's not just you. There seems to be something off in Travis itself.