@seav 2017-11-15
@alukach: I have submitted PR #1894 that I think resolves this Travis error.