@alukach 2017-12-05
Okay, I think we should be good to go.  Based on @oliverroick revelation about needing the permissions filter for the API view, I had to re-add the Tutelary logic to the Project's API view (in fairness, the permissions filter is pretty cool and useful).