@alukach 2017-12-05
I have no idea how the test coverage decreased on the API view file, will resolve shortly...